Move deserialise to session_file
